adjust to renamed BusNotifications
