+++ nservicebus/upgrades/5to6.md
@@ -436,7 +436,8 @@ snippet: 5to6CriticalError
 
 ## Notifications
 
-`BusNotifications` exposed the available notification hooks as observables implementing `IObservable`. This meant a custom `IObserver` and the use of [Reactive-Extensions](https://msdn.microsoft.com/en-au/data/gg577609.aspx) was required to use this API. In Version 6  notifications has been changed to expose regular events instead of observables for easier usage. Find out more about [subscribing to error notifications](/nservicebus/errors/subscribing-to-error-notifications.md). To continue using Reactive-Extensions the events API can be transformed into `IObservable`s like this:
+The `BusNotifications` class has been renamed to `Notifications`.
+`BusNotifications` exposed the available notification hooks as observables implementing `IObservable`. This meant a custom `IObserver` or the use of [Reactive-Extensions](https://msdn.microsoft.com/en-au/data/gg577609.aspx) was required to use this API. In Version 6 notifications have been changed to expose regular events instead of observables for easier usage. Find out more about [subscribing to error notifications](/nservicebus/errors/subscribing-to-error-notifications.md). To continue using Reactive-Extensions the events API can be transformed into `IObservable`s like this:
 
 snippet: ConvertEventToObservable
 
@@ -632,7 +633,7 @@ snippet: 5to6-MsmqSubscriptionAuthorizer
 
 ## Synchronous request-response (callbacks)
 
-The synchronous request-response feature, also known as callbacks, has been moved from the NServiceBus core to the separate Nuget package [NServiceBus.Callbacks](https://www.nuget.org/packages/NServiceBus.Callbacks/). That package must be used in order to use the callback functionality in Version 6. 
+The synchronous request-response feature, also known as callbacks, has been moved from the NServiceBus core to the separate Nuget package [NServiceBus.Callbacks](https://www.nuget.org/packages/NServiceBus.Callbacks/). To use the callback functionality in Version 6, use this new package.
 
 The API was also modified. Version 6 API is asynchronous by default and allows to easily access the response message. It is no longer possible to use callbacks inside handlers or sagas, because extension methods are only available on the message session. The differences in the API are fully covered in [handling responses on the client side](/nservicebus/messaging/handling-responses-on-the-client-side.md).
