+++ samples/logging/notifications/Version_6/Sample/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NServiceBus" version="6.0.0-unstable1705" targetFramework="net452" />
+  <package id="NServiceBus" version="6.0.0-unstable1725" targetFramework="net452" />
 </packages>
